Feature/population disaggregation

+++ app/database/database_functions/data_preparation/walkability/select_weight_walkability_range.sql
@@ -1,5 +1,5 @@
 DROP FUNCTION IF EXISTS select_weight_walkability_range;
-CREATE OR REPLACE FUNCTION select_weight_walkability_range(attribute_input text, condition_input float)
+CREATE OR REPLACE FUNCTION select_weight_walkability_range(attribute_input text, condition_input numeric)
 RETURNS numeric AS
 $$
 	SELECT (w.value*w.weight)
@@ -9,5 +9,3 @@ $$
 	AND (w.max_value >= condition_input or w.max_value IS NULL) 
 $$
 LANGUAGE sql immutable;
-
-
